refactor(core): CryptographyHelper with nullable reference types
* Introduce nullable reference types
* Some clean up
